In mysqldump, skip the table if it's already dropped.
Summary:
mysqldump gets a list of all tables in a database before it dumps any table, then
dumps each table one by one. So there is a possibility that a table is dropped between
mysqldump gets the table list and actually dumps the table. When this happens, the
process maybe die. This diff makes sure that if a table is dropped mysqldump won't crash
but will only log an error. There is still a small window between
  "show table status like ..." and "show create table ..." that if a table is dropped
within this window the process might fail. But this is at least a much smaller window
(sub second vs. up to hours before).

Test Plan:
> create table t (t int); create table u (u int);
> lock table t write, u write;
<-- run mysqldump while holding this connection open -->
> drop table u;
> unlock tables;

in the output from mysqldump (with --skip-lock-tables), you should see:

Error: Couldn't read status information for table u ()
-- Skipping dump data for table 'u', it has no fields

Reviewers: steaphan

Reviewed By: steaphan

CC: ebergen, lachlan, webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1278861

Task ID: 4051194