Remove srv_unzip_LRU_pct and innodb_lru_io_to_unzip_factor
Summary:
I confirmed with perf that we don't actually use this at all in prod,
and haven't for some time.

The test for this frequently fails, and we've spent some time on fixing
it, but the best I got from perf was Yoshi saying we 'maybe' want to use
them in the future for something, but it's ok if we remove them.

That's a pretty solid case for dropping these.  However, I am not going
to factor this back into the history yet, in case we change our mind,
and decide to revert this revert.

Test Plan:
Jenkins

Reviewers: rongrong

Reviewed By: rongrong

CC: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1280049
Differential Revision: https://phabricator.fb.com/D1280108

Task ID: 2291752