Fixing slow log values underflow
Summary: This patch is to fix a problem in the slow log when extra
information is logged. The root cause happens during multi-statement
execution. When in multi-statement, slow log writes the status of
current query, by substracting the last (saved) query status from the
thread's status_var. The problem is, if the thread's status_var is reset
(e.g. FLUSH STATUS) in the middle of the multi-statement, the slow log
will have under-flow (negative value stored in unsigned) values due to
the substraction.

In this fix, we will check if the start status is indeed the true start,
before using any substraction. I tried to avoid making copy of the
status struct to minimize the overhead. Also I made a minor improvement
to check if the query_start and thd->status_var actually points to the
same data to avoid unnecessary substraction for the original code.

Test Plan: Ran slow_log related tests and passed. Also ran arc unit
--everything.

Added a regression test in slow_log_extra.test.

Reviewers: rongrong

Reviewed By: rongrong