Check return values of my_fwrite in mysqlbinlog
Summary:
mysqlbinlog does not check the return value of my_fwrite() when writing
the new data file. The application will keep running despite hitting
ENOSPC/EIO errors without ever reporting a problem. A user running the
program may not realize the bin logs are not being creating correctly.

Check the return values of my_fwrite(), and if they do not match the
expected write bytes, fail the application.

Test Plan:
1. Run mysqlbinlog against a mysqld instance and copy the bin logs.
Repeat with an error injected and verify the application fails with an
error message.

Reviewers: jtolmer, dmituzas

Reviewed By: dmituzas