Increase UNIV_PAGE_SIZE_MAX to 32KB
Summary:
MySQL 5.6 already has support for configurable page size so bumping up
the max is all that is required for this support, unlike in 5.1.

Unlike the other allowable page sizes, the extent size was increased to
2MB in order to keep the descriptor array small enough to fit on a page.
Another option was to remove the unused XDES_CLEAN_BIT from the array but
that solution didn't work for a key block size of 1KB. Also, increasing
the extent size keeps the assert in in buf_dblwr_init that
srv_doublewrite_batch_size is < 2 * TRX_SYS_DOUBLEWRITE_BLOCK_SIZE quiet
so there was no need to reduce the batch size.

Also, at page sizes above 16KB, the offset stored in the dense directory
of compressed pages will exceed the 14 bits allocated for it by default.
Thus, 32KB pages increase the slot size to 3 bytes and give a full 2 bytes
for the offset.

Because UNIV_PAGE_SIZE has a type now ((ulint) srv__page_size) some precompile assertions
in page0zip won't work anymore. Change those to runtime assertions.

Internal:
This is a port of the following changes which initially ported 32k page to 5.6.
https://phabricator.fb.com/D595699
https://phabricator.fb.com/D622580

Test Plan: mtr and mtr --mysqld=--innodb-page-size=32k

Reviewers: jtolmer, steaphan, tianx

Reviewed By: jtolmer, steaphan, tianx

Subscribers: steaphan, webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1337667
Differential Revision: https://phabricator.fb.com/D1546293

Tasks: 4345856, 4910687