relax xtrabackup assert
Summary:
Either relay_log_info_repository=TABLE or gtid_mode=ON
is sufficient on the source instance for the newly created replica
using xtrabackup to be consistent.

xtrabackup currently dies if relay_log_info_repository is not TABLE.
relay_log_info_repository=TABLE is not required if gtid_mode=ON. So,
relax the condition.

Test Plan: dbatool cp_bin on a prod UDB with GTID using file repository.

Reviewers: rongrong, steaphan

Reviewed By: steaphan