Allow to observe binlog cache of foreign threads
Summary:
This patch adds a new SHOW command to output binlog cache content for a
specified thread. The command format is as follows:

SHOW BINLOG CACHE [FOR thread_id] [LIMIT [offset,] row_count]

Notes: the output columns are "Event_type", "Server_id" and "Info". I didn't
add the thread_id in the output, since I think it appearing in the command
makes the output clear. By default (i.e. without thread_id option), it shows
the own binlog cache.

Some implementation notes:

The only lock I grabbed is the foreign thread's LOCK_thd_data to prevent
the thread being deleted. For the binlog cache, I made a copy (snapshot)
of the cache metadata (IO_CACHE), and reset the WRITE_CACHE to
READ_CACHE of the copy. When we read the cache content, we create a
log_event object from the buffer. In this way, the original cache
metadata are not affected and the foreign thread can continue
writing/appending to it. And the original cache buffer is protected by
LOCK_thd_data so that it won't be freed during reading. The side effect
is if the cache content is altered or cleared during reading, we may
read garbage data. That's OK since invalid log_event will not be
outputted.

Test Plan:
Added unit test cases in the new test file "show_binlog_cache.test".

Reviewers: santoshb

Reviewed By: santoshb