Add support to fill factor
Summary:
Added two parameters to control the fill factor of a page:
innodb_defragment_fill_factor: a number between [0.7, 1] that tells the defragment thread
    in what percentage it should try to fill a page. Default is 0.9.
innodb_defragment_fill_factor_n_recs: a number betwee [1, 100] that tells the defragment thread
    it should try to leave space for how many records on the page. Default is 20.

The goal of adding these parameters is to reserve some space on the page during defragmentation so
the next few insert/update on the page will not cause page split. In reality,
  reserved_space = min(page_size * (1 - innodb_defragment_fill_factor),
                       average_record_size * innodb_defragment_fill_factor_n_recs);

The idea is to reserve just enough space to avoid immediate page split. For example, in the default
setting, if the index has larger records so there are less than 200 records on a page,
innodb_defragment_fill_factor will be used to reserve only 10% of the page. If the index has small
records so there are more than 200 records, innodb_defragment_fill_factor_n_recs will be used and
we only reserve enough space for 20 records.

Test Plan:
innodb.innodb_alter_table_defragment is modified to cope the new configuration.
  In the test you can see that since we only fill the page to 90% full when defragmenting, the
  end result of defragmentation is there are about 10% pages. We also modified the test to show
  that a few insertions right after defragmentation will not cause the index size to grow.

Reviewers: pengt, steaphan, jtolmer, santoshb

Reviewed By: santoshb