Do not create temporary tables for UNION ALL
Summary:
mysqld uses a temp table for both UNION and UNION ALL processing. The temp
table for UNION ALL is not needed as results can be immediately returned
to the client. That change would save the cost of writing the temp table
and possibly spilling it to disk. It would also return the first-N rows
to the client much faster.

Test Plan: Jenkins + RPM tests

Reviewers: pengt

Reviewed By: pengt