Prune dead code from xtrabackup.cc
Summary:
The code for xtrabackup was designed to work for a bunch of different
versions of mysql.  These are all either obsolete, or Percona-specific
except for 5.6, which is all we care about here, since we don't maintain
this as a separate diff, but as an integral part of our mysql repository.

So, we don't need all this other stuff.  All of this removed here is
completely dead code.  This should be a non-change.

Internal:
I am pretty confident this is all right, but a close review is worthwhile,
just to be sure.  We all know xtrabackup is super fragile.

Test Plan: Our xtrabackup mtr tests continue to pass.

Reviewers: tianx

Reviewed By: tianx

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1394887

Tasks: 2802340