Revert "Write old FB crcs by default"
Summary:
This change was to ensure that, in a worst-case scenario, we could
always fall back to running 5.1 on the same data.  We no longer need
that, as 5.1 is extinct in all our production environments, and we
are now confident that any non-production stragglers will not need to
downgrade back after being upgraded to 5.6.

So, this removes this, making the written checksums of our branch
fully compatible with the new 5.6 checksums (CRC32 or CRC32_STRICT).

Test Plan:
This change was actually kept as a separate diff for this very purpose.
We have already tested this functionality without this diff.

Reviewers: jtolmer

Reviewed By: jtolmer