Avoid test for innodb_defragment_fill_factor with 32KB pages
Summary:
Since the behavior of these two parameters vary with different page size,
this test is only enabled for 16k pages. For larger page size like 32k,
innodb_defragment_fill_factor_n_recs is more likely to be the one used.
The logic is not related to page size so it should be ok to only test
for 16k page and expect it to not cause problems for other page size.

Test Plan: Jenkins

Reviewers: steaphan

Reviewed By: steaphan