Port ASYNC_COMMIT to 5.6
Summary:
Adds the ASYNC_COMMIT clause for insert, update, delete, replace and
commit statements. The clause can be used at the end of the statement
and when used the InnoDB log is not forced to disk during prepare and
commit steps nor is the binlog forced to disk.

Setting SQL_ASYNC_COMMIT = 1 results in the ASYNC_COMMIT option being on
for all statements.

This also includes the binlog_fsync_count stat.

Test Plan: Test files included.  They pass.

Reviewers: nponnekanti, mcallaghan, steaphan

Reviewed By: steaphan