find binlog by gtid functionality
Summary:
A new query FIND BINLOG GTID='gtid' functionality is given. It prints
the binlog file name and the starting position where the requested gtid
event is physically present. The query reverse iterates over the
previous_gtid_set_map to find the binlog file where the gtid is present
and scans the resulted binlog file comapring the Gtid_log_events with
the requested gtid to find the starting position of the corresponding
Gtid_log_event. A cache of gtid and physical coordinates is helpful to
speed up this procedure, i would like to evaluate this option in another
task.

Test Plan:
mtr test to verify basic functionality of the command
and varous error cases.

Reviewers: jtolmer, steaphan

Reviewed By: jtolmer, steaphan

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D816787
Differential Revision: https://phabricator.fb.com/D1569296

Tasks: 2167563, 5102534