Make mysqld "core_files" sys-var dynamic
Summary:
This diff makes mysqld "core_files" sys-var dynamic.

When this sys-var is true, mysqld will dump core on fatal errors. This sys-var was implemented in two parts as below: the first part, command line option “core-file” is defined, internally its option id is enum OPT_WANT_CORE, when this option “core-file” is in command line arguments or configure file, sys-var Sys_core_file will be set. Here sys-var Sys_core_file, whose type is Sys_var_test_flag, is the second part. This sys-var is a read-only flag defined by bit TEST_CORE_ON_SIGNAL (value is 256) in global uint test_flags, along with a couple of other read-only flags.

To make it dynamic, the type of the sys-var Sys_core_file was changed from Sys_var_test_flag to Sys_var_mybool, it is a regular sys-var that can be set during rum-time, but it is not a command line option, because the original definition of the command line option, which is the first part above, was kept as before. The underlying variable is my_bool opt_core_file, the bit flag  TEST_CORE_ON_SIGNAL was removed.

Test Plan:
The existing test case sys_vars.core_file_basic was enhanced.

Also, because this indicator was read-only but now it is dynamic, we need to make sure it is safe when enabling or disabling it anytime during runtime. Below are all the usages of this sys-var in mysql:

In mysqld.cc, search “opt_core_file”, there are 5 occurrences, basically these are all setting this variable during the initialization when mysqld starts, so it is safe

In function handle_fatal_signal in signal_handler.cc, there is a occurrence. On fatal errors, core will be dumped if it is true. Please note that, in mysql, the updating of sys-var are protected by mutex, but reading usually is not protected. So there is race here, but the read value it either true or false anyway even race happens, the worst case is the core is being dumped when it is set to false, or it is not dumped on fatal errors when it is being set to true, but this also can be treated as timing issue — it is too late when this flag was setting since the fatal error happened already.

So overall, it is safe to change the value anytime during runtime.

Reviewers: steaphan, jtolmer

Originally Reviewed By: steaphan

Reference Patch: https://github.com/facebook/mysql-5.6/commit/4e26365d2bc2