Revert "Disable ibuf and other stuff in xtrabackup"
Summary:
Remove this hack, as it's not a general enough solution.

Internal:
This reverts: https://phabricator.fb.com/D1424905

Squash with https://phabricator.fb.com/D1424905

Test Plan:
This is a clean revert

Internal:
Test copies on ginsights failed a lot without this diff, but the rest
of prod is ok.  Unfortunately this diff triggered problems in other
workloads, so this can't be part of our general solution.

I will continue to include this patch in the custom xtrabackup version
on ginsights until a real fix can be devised.  I've reached out to
Percona for help with that.

Reviewers: ebergen, jtolmer, rongrong

Reviewed By: rongrong

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1447058

Tasks: 4478223
(cherry picked from commit 39ac123b4173bf4e85499e8537fe90c4709c8adf)