Changes in preparation for compact metadata format.
Summary:
Having one function to store metadata of the records on the trailer makes
things complicated especially when the blob pointer storage comes
before the transaction id rollback pointer storage. This is because when
we are compressing a page we do not know how many blob pointers the
page has and in order to allocate room for blob pointers we need this
number. If we compute the number of blob pointers for every page, this might
be inefficient because we know that most of the pages have no blob pointer.
So we do a lazy evaluation: we compute the number of blobs if there is at
least one blob pointer. This diff is a preparation for that change.

For compact metadata format, the blobs are stored before the transaction ids
and rollback pointers on the trailer.

Also for compact metadata format, the number of blobs are stored on the trailer
before the blob pointers themselves.

Test Plan: mtr

Reviewers: rongrong, herman, tianx

Reviewed By: rongrong, tianx