Add error suppression to binlog_write_error test
Summary:
binlog_write_error.test may sometime fail due to the error
'Failed to write the DROP statement for temporary tables to binary log'
printed in the error log.

ER_ERROR_ON_WRITE is expected since write error to binlog is intentionally
injected and this is filtered in the test. But when close_temporary_tables
is called following THD::cleanup(), instead of pushing error to diagnostic area,
the error mentioned above is printed in the log.

Test Plan: arc unit

Reviewers: rongrong, steaphan

Reviewed By: steaphan

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1467392

Tasks: 4797109