Revert "Compress the modification log when it is full."
Summary:
I think the compress modification log change might be problematic.
There are assertion failures in page_zip_size_check, as well as other places
which seem to suggest that an log event of modification log is compressed is
not enough to generate the same compressed image after crash. Reverting the change
and see whether tests still fail.

Test Plan: waiting for jenkins

Reviewers: steaphan

Reviewed By: steaphan