Add support for disabling defrag in stress tests
Summary:
This allows individual stress tests to trun off the defragmentation thread
to see if it changes the behavior.

This is useful for tracking down failures.  But, by default, this does not
change anything.  Defrag remains enabled unless no_defrag is set to 1.

Test Plan:
verified defragmentation log is not present when this is set to 1.

Reviewers: rongrong, steaphan

Reviewed By: steaphan

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1495409
Differential Revision: https://phabricator.fb.com/D1507502

Tasks: 3695297