Better error detection in innodb_stress tests
Summary:
Refactor load_generator.py so that there is a bit less code
duplication among the various worker types. Doing so allows
for consistent error detection so that main can look at the
worker to know if an unhandled expception was encountered.
If one was, fail the test instead of silently ignoring it.

Fix DefragmentWorker to work correctly with crash tests by
handling "Server has gone away" errors and trying again.

The checksum worker and validate_msg now rely on raising
exceptions for error handling.

Once tested and found to be stable, this commit will need to be
divided up and squashed with multiple previous commits, some from
upstream.

Test Plan: Jenkins

Reviewers: hermanlee4, santoshb

Reviewed By: santoshb

Subscribers: webscalesql-eng, tianx, maykov, tengli, pengt

Differential Revision: https://reviews.facebook.net/D48753
Differential Revision: https://reviews.facebook.net/D50067