Add more debugging info for remaining doublewrite buffer test failures
Summary:
This is just a temporary diff to, hopefully, explain why this one case
is failing.  I can't repro this in isolation, not even in Jenkins.

This extra info will allow me to narrow the problem down considerably, the
next time it fails.  This will only output a few extra lines on each crash
recovery, so this should not cause any real log bloat.

This is just a temporary diff.  Once it hits a few times, I will just
remove it.  It will never be public.  So:

@internal

Test Plan:
Built it cleanly.
Jenkins will tell me if this someone breaks a test (it should not).

Reviewers: santoshb

Reviewed By: santoshb

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1525918

Tasks: 4877122