Do not check or try to recover empty doublewrite pages
Summary:
The pages in the doublewrite buffer are fixed.  When the reduced mode is
dynamically enabled, only the first page may be used, so the other pages
are overwritten with blank pages that point to themselves.

Since these pages *may* be used by the (separate) single-page flushing
system, we still have to process them.  The idea was that these pages
would pass the checksums, and so nothing special had to be done.

There is really no point in checking a doublewrite page that points
within the doublewrite buffer itself.  This was just done to keep the
code simple.  So, now, we just ignore these pages entirely.

Internal:
The implementation included in fb-mysql-5.6.12 created these pages'
checksums incorrectly.  So, any unused pages left over from then will
actually fail to recover, since both the doublewrite page, and the page
it points to (which are both the same page) fail to verify.  This is
actually a common situation on flash, since we flush so frequently, most
of the doublewrite buffer remains unused most of the time, so most pages
are left in this initial unused state.

Test Plan:
Jenkins

mysqltest.sh --clean --release --repeat=16 --parallel=16 --mem \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite

Verified using extra logging info that these pages (and *only* these
pages) are being skipped during all this testing.

Reviewers: pengt

Reviewed By: pengt

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1525918
Differential Revision: https://phabricator.fb.com/D1652580
Differential Revision: https://phabricator.fb.com/D1666898

Tasks: 5509074