Do not check or try to recover empty doublewrite pages
Summary:
The pages in the doublewrite buffer are fixed.  When the reduced mode is
dynamically enabled, only the first page may be used, so the other pages
are overwritten with blank pages that point to themselves.

Since these pages *may* be used by the (separate) single-page flushing
system, we still have to process them.  The idea was that these pages
would pass the checksums, and so nothing special had to be done.

There is really no point in checking a doublewrite page that points
within the doublewrite buffer itself.  This was just done to keep the
code simple.  So, now, we just ignore these pages entirely.

Test Plan:
Jenkins

mysqltest.sh --clean --release --repeat=16 --parallel=16 --mem \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite \
	innodb.innodb_doublewrite innodb.innodb_doublewrite

Verified using extra logging info that these pages (and *only* these
pages) are being skipped during all this testing.

Reviewers: pengt

Reviewed By: pengt