Fixing Jenkins Test Failure innodb.innodb_alter_table_defragment
Summary:
We have an async_commit defrag on a table that will be dropped soon at
the end of the test. However, if async_commit defragmentation doesn't
finish before drop table, it will cause crash after the table is
dropped.

The purpose of async defrag is to see if it will block another defrag on
the same index. There's actually no need to wait for it to finish. So we
should clean up the unfinished query as part of the test cleanup.

Test Plan: innodb.innodb_alter_table_defragment

Reviewers: steaphan, rongrong

Reviewed By: rongrong

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1534183

Tasks: 4958700