Use read_only for GTID deployment
Summary:
Slaves should not generate GTIDs on it's own in the slave binlog. This
can break replication with GTID protocol when the slave which generated
it's own GTIDs is promoted as the master. All slaves in the replication
tier can fail with SQL_thread error 1236 after promotion with message
"The slave is connecting using CHANGE MASTER TO MASTER_AUTO_POSITION =
1, but the master has purged binary logs containing GTIDs that the slave
requires".

To avoid this slaves should not generate it's own GTIDs but should use
GTIDs received from master. This functionality is provided by read_only
which is 1 on slave and 0 on master.

Now, with read_only enabled, a host cannot generate gtids on its own,
but if gtid logged events are received through replication stream from
master, gtids will be logged.

Modified logic such that sql_thread never generates gtids automatically
on it's own.

Test Plan:
Tested replication is working with various gtid deployment scenarios.

Reviewers: jtolmer

Reviewed By: jtolmer

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D761724
Differential Revision: https://phabricator.fb.com/D962493

Task ID: 2838624

Conflicts:
	sql/rpl_master.cc