Add mysqldump option to print ordering key
Summary:
The --print-ordering-key option to mysqldump causes mysqldump to get the
key which is used to order the rows for each table, and print it to the
dump file.

This will help to efficiently run a diff on two mysqldumps of the same
database at different points of time.

Test Plan:
Ran mysqldump with this option. Sample output snippet:

CREATE TABLE `xid_map` (
  `fbid` bigint(20) unsigned NOT NULL DEFAULT '0',
  `fbtype` int(10) unsigned NOT NULL DEFAULT '0',
  `xid` varchar(255) NOT NULL DEFAULT '',
  PRIMARY KEY (`fbid`),
  UNIQUE KEY `fbtype_xid` (`fbtype`,`xid`)
) ENGINE=InnoDB DEFAULT CHARSET=latin1;
/*!40101 SET character_set_client = @saved_cs_client */;
/* ORDERING KEY : fbtype_xid */;

LOCK TABLES `xid_map` WRITE;
/*!40000 ALTER TABLE `xid_map` DISABLE KEYS */;
INSERT INTO `xid_map` VALUES  .....

Checked that the rows are actually sorted with respect to printed key.

Reviewers: steaphan

Reviewed By: steaphan

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D880160

Task ID: 2577764

Conflicts:
	client/client_priv.h