Fix master hang during shutdown
Summary:
We are seeing cases of masters hanging during shutdown because bin log
dump threads are sitting waiting for more bin log events to get generated,
which will never come. This problem is one that I've seen in the past.
Fix by merging the fix I made in the Google 5.1 patch with the 5.6 version
of kill_zombie_dump_threads to 5.6ify it.

Test Plan: Jenkins

Reviewers: santoshb

Reviewed By: santoshb