flush result file after my_fwrite()
Summary:
mysqlbinlog writes to the file using my_fwrite() which
calls ftell() and then calls fwrite(). ftell() causes the output
buffer to get written to the file. But after doing fwrite() the
recent event is not flushed to the file until another event is received
by mysqlbinlog or when mysqlbinlog process is terminated gracefully.

This causes bug mentioned in http://bugs.mysql.com/bug.php?id=69568.
Fix this by calling fflush() after my_fwrite()in raw_mode when the current
event is XID_EVENT or when mysqlbinlog haven't flushed the result file
even after receving --flush-result-file events from the
server.

Internal:
Note that this diff doesn't care about DDL queries
which are rare to be in binlog because of OSC.

Test Plan:
run mysqlbinlog --raw and check the most recent xid event
is actually written to the result file.
run mysqlbinlog --flush-result-file=3 to check the log is flushed
after receving 3 events (none of them are xid events) from the server

Reviewers: steaphan, jtolmer, yoshinori

Reviewed By: steaphan

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D968422

Task ID: 2688103

Conflicts:
	client/client_priv.h