Add linux thread id to show processlist
Summary:
When running 'show processlist' on linux OS, additionally show system
thread id. It is helpful to have the system thread pid of a connection
to trace individual threads or change process priority.

Previously, we had to look through quickstack output to find the thread
id based on a specific stack.

Also, there is a race in show processlist, which shows the actual
show processlist run as either "init" or "cleaning up".  To avoid
the test inconsistency, I replace both with "STATE" in many tests.

Test Plan: mtr

Reviewers: steaphan, ebergen, jtolmer, santoshb

Reviewed By: steaphan, santoshb