is_in_group() function should be valid with GTID
Summary:
is_in_group() function should made work with GTID

group position is updated by slave sql_thread if it is not currently
executing a group and reads a ROTATE event from the realy log.

But with GTIDs, group position is updated by slave sql_thread
when it reads a rotate event from the relay log and
even after executing a GTID_LOG_EVENT. This causes slave to
miss the GTID if it is stopped exactly after executing the rotate event.
Fix this by updating is_in_group() function to be compatible with
GTID.

Test Plan:
* Change the block size of max_relay_log_size to 1 and set it to a value 870 on a slave
* stop slave
* generate insert into t1 values(1); load on master table
* change master to master_auto_position=1
* start slave sql_thread until relay_log_file='relay_log_2', relay_log_pos=4
* start slave and verify the slave's gtid_executed set. Also verified the
contents of table t1 is same on both master and slave

added a mtr test to consistently reproduce this.

Reviewers: jtolmer

Reviewed By: jtolmer

CC: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1273628

Task ID: 4054112

Conflicts:
	sql/rpl_rli.h