Add thread-id filter option to mysqlbinlog executable
Summary:
This diff adds a "--thread-id" option to the mysqlbinlog tool that
lets you limit the output of queries to the given thread. Non-query
log evetns are still displayed.

Test Plan:
You can pipe the output of mysqlbinlog through 'grep thread_id='
to see what thread_ids are contained in the output. If you want
to see how many queries of any thread_id are contained in the output,
run:
  mysqlbinlog FILENAME | grep "thread_id=" | cut -f 3 | sort -V |
      uniq --count

If you re-run this command line with a "--thread-id=XXX" option
added to mysqlbinlog (where XXX is one of thread-ids appearing
in the binlog file used and output by the first run), you should
only get that one thread_id reported in the output, and the count
should be identical to what was reported previously.

Reviewers: ebergen, santoshb

Reviewed By: santoshb