Move hexdump helper function from page0zip to ut0dbg and improve
Summary:
There is a helper function in page0zip.cc that dumps a memory region as
hexadecimals to stderr. It is going to become useful in other places
as well, so this diff renames it, moves it to ut0dbg.cc and adds a
declaration into ut0dbg.h so that it is usable everywhere. This diff
also changes the formatting of the hexdump, making it look like the
output of "hexdump -C" which should be nice and helpful.

Test Plan:
It compiles. I've looked at a couple of dumps. This is a debug helper
function, forgive me for not adding unit tests, as it is also not
written in a well-testable way.

Reviewers: rongrong, steaphan

Reviewed By: steaphan