More dbg output when assertion in btr_pcur_move_to_next_page fails
Summary:
We see this assertion occasionally fail in the innodb_stress_lra
test. In order to allow debugging, this diff prints additional
information to stderr.

@internal

Test Plan:
I've short circuited the 'if' to be able to see the debug output
once. That's manual testing and should be good enough for this.
I assume this commit is also temporary and can be reverted once
the actual problem could be solved.

Reviewers: steaphan, rongrong

Reviewed By: rongrong

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1552444

Tasks: 5126498