Port Compression: Making empty space reserved dynamic to 5.6
Summary:
* segment_reserve_factor determines the ratio of empty pages, the last
segment in an ibd can have before innodb creates a new extent. If a new
page is needed and the last segment has less than
    segment_reserve_factor * (current number of pages in this segment)
empty pages, then a new extent is created and the new page is allocated
from the new extent. The empty pages are useful when B-tree grows so
that pages in the upper level of the B-tree can be allocated
contiguously. An extent is 64 pages, and a segment is 256 extents.

Before this diff, this variable was not configurable and it was
determined by 1/x where x was the value defined by the compile time
than 1/8th pages were empty, then InnoDB would allocate a new extent
leaving many empty pages in the current extent. This caused lots of
fragmentation.

The ideal value for this variable needs investigation. The current
default will have the previous behavior. A segment has 16384 pages, so
for eg. If we are okay with leaving at most 16 empty pages per segment,
then we can set innodb_segment_reserve_factor=0.001.

innodb_segment_reserve_factor.test loads data into two identical tables
but sets a lower segment_reserve_factor before loading data to the
second table. This makes the second table about 10% smaller than the
first table.

The default value for this variable is set to 0.01 which should mean 1%
of the pages will be allocated for growth.

Internal:
This is a port of the following diffs:

1) D426799 Make Segment Fill Factor a Dynamic Variable.

2) D441063 Fix innodb_segment_reserve_factor output

3) D525505 Change the declared value for innodb_fseg_reserve

The command for comparing ibd file sizes printed the current working
directory when run during rpm build from what I understand.

The default value for innodb_fseg_reserve is 0.01 but its declared as
0.125. Change this to follow the convention.

Test Plan: Tested on Jenkins.

Reviewers: nizamordulu, jtolmer

Reviewed By: nizamordulu

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D747613

Task ID: 2002320

Conflicts:
	storage/innobase/handler/ha_innodb.cc

Conflicts:
	storage/innobase/handler/ha_innodb.cc