xtrabackup: set the upper 32 bits of offset in checkpoint header
Summary:
The checkpoint sections of the redo log file header contain
the 64-bit offset into the log file as two non-contiguous 32-bit
fields. xtrabackup overwrites the checkpoint information of
redo log files and makes the offsets point to the beginning
of the xtrabackup log file (in the range 0x800 to 0x9ff).
However, the upper 32 bits of the offset (which is always 0)
weren't written. In situations where log files exceed 4GB in
size, the upper-32-bit field can originally contain a non-zero
number, so this field must be set to 0 in order to make the
offset point to the beginning of the log file as intended.

Test Plan:
Existing tests must not fail. Manual testing with large log
files will be required.

Reviewers: steaphan

Reviewed By: steaphan

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1568984

Tasks: 4671138