Add missing statement to the innodb_change_zlib* stress tests
Summary:
There was a "disable_query_log" without an "enable_query_log", which led
to a later (correct) change adding output that was not recorded, because
this caused these to be nested (which doesn't work).

I added the missing enable, and re-recorded these tests.

Internal:
I will squash this with the diff that introduces these tests.

Test Plan: Jenkins

Reviewers: rongrong, jtolmer, santoshb

Reviewed By: santoshb

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1575664

Tasks: 5232285