Reduce the race condition in async defragment
Summary:
We check whether an item is removed from defrag list before rate
limiting in defragment thread. This leaves a large window for other
thread to decide to kill the query or remove the table which will
crash the following pcur open. Reorder the code to make that less likely
to happen.

Test Plan: hopefully the crash in innodb.innodb_alter_table_defragment will stop.

Reviewers: tianx, over, steaphan

Reviewed By: steaphan

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1594297

Tasks: 5048831