xtrabackup: make "no background threads" a command line option
Summary:
Internal:
While we are investigating whether the earlier commit (which
disabled background threads in xtrabackup) has harmful side effects
for databases with compressed tables, I have prepared this
diff which would turn the disabling of background threads a
command line option in xtrabackup.
If it turns out we don't need this, please reject this diff.
If we do need it, here it is.

Test Plan: Manual testing.

Reviewers: steaphan

Reviewed By: steaphan

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1591805

Tasks: 5299147