Make sure defragmentation don't block super read only
Summary:
defragmentation is not changing the table in any logical sense so it should not
hold the mdl lock that would block super read only. Set the flag correctly.

Test Plan: added a new test to make sure super read only is not blocked. test would deadlock withouth this change.

Reviewers: steaphan, over

Reviewed By: over