Remove mysql client warning about passwords
Summary:
This warning annoys people, and it's pretty stupid since it just makes
you do it in even less secure ways (like env vars).

This effectively re-includes: https://phabricator.fb.com/D885466

However, every single file in that diff has conflicts, and there are
other changes needed beyond it... so I completely re-wrote this.

We'll remove this diff once we switch to a real creds system.
Meanwhile, we don't want to tell the world we are clowny enough with
security to pass passwords on the command-line, so:

@internal

Test Plan:
Jenkins

And, this shows nothing now (showed many before):
git grep '^Warning: Using a password on the command line interface can be insecure.$'

Reviewers: jtolmer, chip

Reviewed By: chip

Subscribers: webscalesql-eng@

Differential Revision: https://phabricator.fb.com/D1630793

Tasks: 2220881