remove implicit dependency on my_global.h from mysql.h
Summary:
By using the C_MODE_BEGIN/C_MODE_END macros we include an implicit dependency on my_global.h.  This is different than the Oracle mysql.  This breaks the open source hhvm.  The fbcode version is using the 5.6.12 headers that don't have this impicit dependency.

Despite requiring the mysql.h.pp change this is not an ABI breaking change.  Further evidence of this is the HHVM build with 5.6.12 headers works fine with the mysql 5.6.21.

This diff https://reviews.facebook.net/D17019 has the change.  Interestingly the diff originally did not use C_MODE_BEGIN/C_MODE_END but darnaut pointed out the existance of the macro.

The alternative of including my_global.h everywhere was looked at but it has a few issues.  1) is my_global.h pulls in my_config.h which introduces serveral macros that conflict with HHVM.  2) If our mysql is intended to be open sourced and compete with Oracle, we may want to be compatible.

Test Plan: I tried are unit but 8 hours later it still says: Waiting for build to start...

Reviewers: steaphan

Reviewed By: steaphan