Remove redundant test
Summary:
When Oracle gave us the patch for the test case for
public bug 69122, it added the test under a different name
than what they pushed to the official source. We now have the
correct upstream test in innodb-change-buffer-recovery.test.

This diff fully reverts the original diff which added this test
as well as subsequent changes to those files.

Test Plan: None

Reviewers: steaphan

Reviewed By: steaphan