Revert "Skip purged records when serializing except for the record offsets."
Summary:
Revert the change that skips the purge record so we can refactor
compression code easier. One of the goals for refactoring is to get rid
of the streamd zlib compression all together. To do this, we need to
keep the buffered zlib compression and streamed zlib compression format
compatible. However, current buffered zlib compression is only available
with format change. Skiping purged records is one of the things that
changes the data format. So we revert this first.

Test Plan: jenkins

Reviewers: over

Reviewed By: over