Port FBSON library's unit test code
Summary:
This is to port all unit test code (using GTest) into MySQL's fbson lib
directory. The unit tests are not intented to be ran with MySQL's test
framework, but are ported as a referenece for people who wish to use and add
tests to the FBSON library.

Test Plan: FBSON unit tests

Reviewers: pengt

Reviewed By: pengt