Make mysqldump not lock all tables with --single-transaction
Summary:
Change the behavior of --single-transaction to use

	START TRANSACTION WITH CONSISTENT INNODB SNAPSHOT

to begin transactions, which returns the binlog file and offset. This
is later used if --master-data is specified and becomes part of the
output (otherwise it just avoids the lock tables).

Internal:
This is a port of https://phabricator.fb.com/D454218 to 5.6:
Make mysqldump not lock all tables when --single-transaction is used

Test Plan:
mtr tests pass.
main.mysqldump contains a test for this, check for Bug#12809202
Not sure if more testing is required.

Reviewers: jtolmer, steaphan

Reviewed By: steaphan

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D774484

Task ID: 1820981

Conflicts:
	client/mysqldump.c