Add functions to retrieve and set SSL sessions for connections
Summary:
SSL allows for sessions to be re-used, which significantly
improves the time it takes to create connections (~4-5x more connections
per second, with less CPU use).  This requires the ability to retrieve
and set the SSL session (which is an opaque blob of bytes); the client
presents the session to the server which decides whether to re-use it or
not.

We set the session via mysql_options4, and create a new function to
retrieve it from an existing connection.

This also adds a very simple cache to the 'mysql' command line tool --
it will try to re-use the most recently seen SSL session when
re-establishing connections.

It is expected client libraries will manage a cache themselves; nothing
in this diff caches the sessions themselves, instead allowing this
policy to be decided by the application using the MySQL client library.
Typically this would be done with a simple hash map of "host:port" to
the most recent session ID.

Test Plan:
jenkins, plus testing via the command line tool to see
connection speed improvements

Reviewers: santoshb, anca

Reviewed By: anca