Add simple function to view server subject and issuer, and expose via cli
Summary:
This info was previously only available in debug mode trace
files.  Now it is available via an API (sketchy as it may be) as well as
in the mysql command line client when using the '\s' command.  Sample
output:

https://phabricator.fb.com/P19023234

Test Plan: ran mysql cli, saw the server certs via \s

Reviewers: anca, santoshb

Reviewed By: santoshb