Add column names to rbr table map log events.
Summary:
Column names added to table map log events are used by slave
to find out the corresponding field in slave's table whose index
may be different than the index of the field in master's table
due to a schema change.

This behavior is controlled for each table using
  Alter table rbr_column_names=1

This is an inplace alter which does metadata only change in the table
frm file.

During a schema change on a replica set, slaves may have different schema
than master which will break RBR. The following functionality needs to be
added to online schema change (OSC):

a. Before starting OSC, run alter table rbr_column_names=1 on master of the
replica set. This will get replicated to the slaves.

b. Run OSC on all slaves first and then on master.

c. Run alter table rbr_column_names=0 on master of the replica set to
turn off the feature.

The column names will increase the binlog size, so they should be used
temporarily on per-table basis only while OSC is running on the replica set.
This change is backward compatible but older MySQL versions will not understand
the new format of table map log events. So, before using this feature
all the servers in the replica set must be upgraded to new MySQL with this
change.

Note mysqlbinlog is not updated to show table column names when available.

Test Plan: mtr tests

Reviewers: ebergen, jtolmer

Reviewed By: jtolmer

Differential Revision: https://reviews.facebook.net/D52635
Differential Revision: https://reviews.facebook.net/D52797