Add facility to re-use SSL_CTX across connections
Summary:
Every connection, MySQL recreates the SSL_CTX object.  This is
bad and unnecessary.  This diff adds the ability to re-use an SSL
context object between connections.  This is roughly a 50% increase in
throughput making SSL connections.  In combination with re-using SSL
sessions, the net speedup is roughly 7-8x.

This facility can be used in one of two ways; one is to "take ownership"
of the SSL_CTX in an existing MySQL connection (simple) or to create a
context by hand, therefore bypassing many of the MySQL "options" fields
relating to verification, certs, etc.  This is not ideal but avoids a
larger refactoring and public API change to support some kind of
"initialize an OpenSSL context with these options" which would be
complicated and potentially difficult to reconcile with yassl.

Test Plan: runtests, also timed with 'mysql' command line tool
- added ssl-session-reuse test

- manual testing with 'mysql' command line tool:

  build-5.6-Debug/client/mysql --ssl=1 --ssl-ca=ca-cert.pem --ssl-cert=client.cert --ssl-key=client.key -h<db1> -P<port>

  (user@db1:port) -> connect  // reconnect, only SSL_CTX will be set

  (user@db1:port) -> status    // check  ssl still being used

  (user@db1:port) -> show status like 'ssl_session%'  // check session
  reused

  (user@db1:port) -> connect mysql <db2>  // connect to a different db,
  only SSL_CTX will be set

  (user@db2:port) -> status   // ssl connection succeeded

Reviewers: anca, santoshb

Reviewed By: anca, santoshb