Fix MySQL openssl initialization to be thread safe
Summary:
The old method wasn't thread safe as it didn't use a lock.
This isn't an issue for the server when connecting for replication, but
in a client, if two SSL connections are being created at the same time,
then there is very small race on the booleans used to track
initialization state.  This replaces the questionable code with an
actual mutex.  No performance degradation, even in a tight connect loop.

Test Plan: runtests

Reviewers: santoshb

Reviewed By: santoshb