Add mysqldump support for logical read ahead
Summary:
Adds options to mysqldump:
	--lra-size=X
	--lra-sleep=X
	--lra-n-node-recs-before-sleep=X

These just inject SET statements to set these session variables.

Test Plan:
Using a previous version of this which triggered an error if LRA was
not present in the server, not just a warning (as this version does):

Ran the mysqldump test on mysqld WITHOUT logical_read_ahead:

--lra-size=5
	Failed on "SET innodb_lra_size=5" with unknown variable:

	mysqldump: Couldn't execute 'SET innodb_lra_size=5':
	Unknown system variable 'innodb_lra_size' (1193)

--lra-size=5 --lra-sleep=10
	Failed on "SET innodb_lra_size=5" with unknown variable:

	mysqldump: Couldn't execute 'SET innodb_lra_size=5':
	Unknown system variable 'innodb_lra_size' (1193)

--lra-sleep=10
	Worked (passed nothing) - forced the code to send anyway, and got:
	mysqldump: Couldn't execute 'SET innodb_lra_sleep=10':
	Unknown system variable 'innodb_lra_sleep' (1193)

	...and likewise for --lra-n-node-recs-before-sleep=X

Ran all of this on mysqld WITH logical_read_ahead:
	Tests all passed without error.

Using this version, which only generates a warning:

Ran mysqldump --lra-size=8 on a table on 5.1 and 5.6 servers.
On 5.1, printed the warning, and dumped normally.
On 5.6, no warning, and dumped normally.

Reviewers: nizamordulu

Reviewed By: tianx
(Original Reviewed By: nizamordulu)

CC: yoshinori, lachlan, mysql-eng@

Differential Revision: https://phabricator.fb.com/D769460
Differential Revision: https://phabricator.fb.com/D932796

Task ID: 2273538, 2526544

Conflicts:
	client/client_priv.h

Conflicts:
	client/client_priv.h
	client/mysqldump.c

Conflicts:
	client/client_priv.h