Stablize main.slow_log_extra
Summary:
An attempt to stabilize this intermittenly failing test.

Change the test to initially start with a very long long-query-time so
that we don't get unexpected, additional log entries on heavily loaded
test hosts.

Remove the initial sanity check that depends on a sleep as that basic
funcionality testing is redundant with all the additional work that's
since been added to the test.

Test Plan:
mtr --parallel=32 --repeat=32 main.slow_log_extra ... is clean on my devserver
with this diff. Ultimately, Jenkins nightly runs will tell us if it's
really fixed.

Reviewers: santoshb

Reviewed By: santoshb