[DocStore] Maintain secondary indexes on document paths
Summary:
The following new columns are added to SYS_FIELDS table which store metadata information required to
support indexing on document paths

  Document path  |  Document path type

When creating a new index on document path, new entries are made to the persistent SYS_FIELDS table.
The first columne stores the full document path and the second column stores the type information of the
field. Dictionary cache is enhanced to understand the new metadata information.

On every DML operation, if the index contain a document path, the corresponding value is extracted
from the binary data which is used in updating the index. The following is the mapping from Fbson
types to Innodb types:

T_Null                         NULL
T_True | T_False               DATA_INT with size 1
T_Int*                         DATA_INT with size 8
T_Double                       DATA_DOUBLE
T_String                       DATA_BINARY
T_Binary | T_Object | T_Array  DATA_BLOB

If the type conversion is not possible which depends on the type specified when creating the index, simply
NULL is stored. A NULL is also stored in the index if document path type is FbsonType::T_Null or if the document
path value is not found in the document field. Note that innodb allows multiple NULL values even for unique
indexes, so each row insert will have an index entry.

Test Plan:
mtr
Verify the new mysqld understands old SYS_FIELDS format

Reviewers: tianx, pengt

Reviewed By: pengt